Rename btree directory to blocks.
